+++ test_urls.py
@@ -1,1 +1,3 @@
-urlpatterns = []
+from django.urls import include, path
+
+urlpatterns = [path("rest/", include("email_auth.interfaces.rest.urls"))]