+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ Features
 * :issue:`32`: Link to frontend pages for resetting passwords and verifying
   email addresses. These links are specified with the ``PASSWORD_RESET_URL`` and
   ``EMAIL_VERIFICATION_URL`` settings, respectively.
+* :issue:`36`: Add endpoint to verify an email address to the provided REST
+  interface.
 
 ******
 v0.3.1